Crash Fix and StatsComponent
Fixed a crash related to a nullptr when trying to receive the current navigation system v1 with a UObject, instead it will use an actor component. I don't know why it doesn't work with a UObject (since it does have GetWorld() and the world context ist needed) but it is what it is
Added a Stats Component and more GameplayTags to use with the stats component. Stats will be managed through it